Типизирован слайс total-price
